+++ subprojects/core-impl/src/main/groovy/org/gradle/api/internal/artifacts/ivyservice/modulecache/ModuleDescriptorCache.java
@@ -20,6 +20,8 @@ import org.apache.ivy.core.module.id.ModuleRevisionId;
 import org.gradle.api.artifacts.ResolvedModuleVersion;
 import org.gradle.api.internal.artifacts.ivyservice.ivyresolve.ModuleVersionRepository;
 
+import java.math.BigInteger;
+
 public interface ModuleDescriptorCache {
     void cacheModuleDescriptor(ModuleVersionRepository repository, ModuleRevisionId resolvedModuleVersionId, ModuleDescriptor moduleDescriptor, boolean isChanging);
 
@@ -34,7 +36,7 @@ public interface ModuleDescriptorCache {
 
         long getAgeMillis();
 
-        int getDescriptorHash();
+        BigInteger getDescriptorHash();
 
         boolean isMissing();
     }
